requirements.txt
autopep8==1.6.0
colorama==0.4.4
commonmark==0.9.1
mypy==0.910
mypy-extensions==0.4.3
pycodestyle==2.8.0
Pygments==2.10.0
rich==10.14.0
toml==0.10.2
typing-extensions==4.0.0
vunit-hdl==4.6.0
